remove test transpose
